SQUASH: Hoist ResultCollection up to qiime2 not just qiime2.sdk
