+++ qiime2/__init__.py
@@ -6,7 +6,7 @@
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
 
-from qiime2.sdk import Artifact, Visualization
+from qiime2.sdk import Artifact, Visualization, ResultCollection
 from qiime2.metadata import (Metadata, MetadataColumn,
                              CategoricalMetadataColumn, NumericMetadataColumn)
 from qiime2.plugin import Citations
@@ -23,9 +23,9 @@ __release__ = '.'.join(__version__.split('.')[:2])
 __citations__ = tuple(Citations.load('citations.bib', package='qiime2'))
 __website__ = 'https://qiime2.org'
 
-__all__ = ['Artifact', 'Visualization', 'Metadata', 'MetadataColumn',
-           'CategoricalMetadataColumn', 'NumericMetadataColumn', 'Cache',
-           'Pool']
+__all__ = ['Artifact', 'Visualization', 'ResultCollection', 'Metadata',
+           'MetadataColumn', 'CategoricalMetadataColumn',
+           'NumericMetadataColumn', 'Cache', 'Pool']
 
 
 # Used by `jupyter serverextension enable`
