+++ qiime2/core/type/collection.py
@@ -80,7 +80,7 @@ class _List(_1DCollectionBase):
         """Since this is a dictionary, we often need to make sure to use its
         values not its keys.
         """
-        from qiime2.sdk import ResultCollection
+        from qiime2 import ResultCollection
 
         contained_expr = self_expr.fields[0]
 
@@ -120,7 +120,7 @@ class _Collection(_1DCollectionBase):
         """Since this is a dictionary, we often need to make sure to use its
         values not its keys.
         """
-        from qiime2.sdk import ResultCollection
+        from qiime2 import ResultCollection
 
         contained_expr = self_expr.fields[0]
